+++ lib/picons.dart
@@ -3,3 +3,4 @@ library picons;
 export 'package:picons/src/picon_data.dart';
 export 'package:picons/src/picon.dart';
 export 'package:picons/src/picons.dart';
+export 'package:picons/src/picons_theme.dart';